+++ src/components/location-finder/Location/LocationHours.test.js
@@ -69,7 +69,7 @@ describe('LocationHours Component', () => {
       />
     );
 
-    expect(screen.getByText('Closed.')).toBeInTheDocument();
+    expect(screen.getByText('Closed')).toBeInTheDocument();
   });
 
   test('location with regular hours should be temporarily closed due to extended closing or alert closing.', () => {
@@ -105,7 +105,7 @@ describe('LocationHours Component', () => {
       />
     );
 
-    expect(screen.getByText('Closed.')).toBeInTheDocument();
+    expect(screen.getByText('Closed')).toBeInTheDocument();
   });
 
   test('location with null for end hours only should be closed due to regular hours', () => {
@@ -117,7 +117,37 @@ describe('LocationHours Component', () => {
       />
     );
 
-    expect(screen.getByText('Closed.')).toBeInTheDocument();
+    expect(screen.getByText('Closed')).toBeInTheDocument();
+  });
+
+  test('Location with regular hours and by appointment only should render hours notice.', () => {
+    render(
+      <LocationHours
+        open={true}
+        todayHoursStart={'11:15'}
+        todayHoursEnd={'14:01'}
+        appointmentOnly={true}
+      />
+    );
+
+    expect(screen.getByText('* Division is by appointment only.')).toBeInTheDocument();
+    expect(screen.getByText("Today's Hours:")).toBeInTheDocument();
+    expect(screen.getByText('11:15AM–2:01PM*')).toBeInTheDocument();
+  });
+
+  test('Location with regular hours, but not by appointment only should not render hours notice.', () => {
+    render(
+      <LocationHours
+        open={true}
+        todayHoursStart={'11:15'}
+        todayHoursEnd={'14:00'}
+        appointmentOnly={false}
+      />
+    );
+
+    expect(screen.queryByText('* Division is by appointment only.')).not.toBeInTheDocument();
+    expect(screen.getByText("Today's Hours:")).toBeInTheDocument();
+    expect(screen.getByText('11:15AM–2PM')).toBeInTheDocument();
   });
 
   // Accessbiility tests.