+++ src/components/shared/Dropdown/Dropdown.js
@@ -1,0 +1,60 @@
+import React, { useState } from 'react';
+import { Icon } from '@nypl/design-system-react-components';
+import FocusTrap from 'focus-trap-react';
+
+function Dropdown(props) {
+  const {
+    id,
+    label,
+    children,
+    checked,
+    onChange,
+    hasSelectedItems
+  } = props;
+
+  return (
+    <div className={hasSelectedItems 
+      ? `dropdown hasSelectedItems` 
+      : `dropdown`}
+    >
+      <FocusTrap
+        focusTrapOptions={{
+          clickOutsideDeactivates: true,
+          returnFocusOnDeactivate: false,
+        }}
+        active={checked}
+      >
+        <div>
+          <input
+            id={`dropdown-${id}`}
+            type="checkbox"
+            className="dropdown__input"
+            checked={checked}
+            onChange={onChange}
+          />
+          <label
+            htmlFor={`dropdown-${id}`}
+            className="dropdown__label"
+          >
+            <span className="dropdown-label">{label}</span>
+            <Icon
+              decorative={true}
+              name="minus"
+              modifiers={["small", "minus"]}
+            />
+            <Icon
+              decorative={true}
+              name="plus"
+              modifiers={["small", "plus"]}
+            />
+          </label>
+          <div className="dropdown__content">
+            {children}
+          </div>
+        </div>
+      </FocusTrap>
+    </div>
+  );
+};
+
+export default Dropdown;