+++ testHelper/__mocks/refineryLocationsMock.js
@@ -1,0 +1,40 @@
+const locations = [
+  {
+    slug: "115th-street",
+    name: "115th Street Library",
+    geolocation: {
+      type: "Point",
+      coordinates: [
+        -73.9532,
+        40.8028
+      ]
+    },
+    contacts: {
+      phone: "(212) 666-9393",
+      manager: {
+        fn: "Tequila Davis",
+        title: "Library Manager"
+      }
+    }
+  },
+  {
+    slug: "125th-street",
+    name: "125th Street Library",
+    geolocation: {
+      type: "Point",
+      coordinates: [
+        -73.9532,
+        40.8028
+      ]
+    },
+    contacts: {
+      phone: "(212) 666-9393",
+      manager: {
+        fn: "Tequila Davis",
+        title: "Library Manager"
+      }
+    }
+  }
+];
+
+export { locations as default };